Create 230614_WellPlateDispenser_v4.py
A script with a simple GUI to generate gcode for dispensing into multi well plates. Input should be the syringe size used, the amount to be extruded in each well, the pause in seconds after the extrusion, the number of wells to print in in each row, and the Z-lift after extrusion before moving to the next well.

Does not work well for uneven numbers of rows.